+++ Changelog.markdown
@@ -91,6 +91,7 @@ There are several categories of change:
   shut down scheduler
 - fix: duplicated names are not allowed in a single source file
 - bic: processes cannot be suspended, and their priority cannot be adjusted, from user code; VM is the sole ruler of processes
+- bic: rename `enclose` family of instructions (`enclose`, `enclosecopy`, and `enclosemove`) as `capture` family
 
 One limitation of static analyser (SA) introduced in this release is its inability to handle backwards jumps.
 This, however, is not a problem if the code does not use loops and